Moving back to code that didn't need to be refactored out
